requirements.txt
flask>=3.0.0
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0
pyserial>=3.5
scipy>=1.11.0
rembg>=2.0.50
onnxruntime>=1.16.0
requests>=2.31.0
pooch>=1.8.0
mediapipe>=0.10.0
